refactor(proto): split `executor/smart_function.rs` into multiple files
