--- crates/jstz_proto/src/executor/smart_function.rs
@@ -1,1707 +1,0 @@
-use std::{num::NonZeroU64, ops::BitXor};
-
-use boa_engine::{
-    js_string,
-    object::{builtins::JsPromise, ErasedObject, FunctionObjectBuilder},
-    Context, JsArgs, JsError, JsNativeError, JsResult, JsValue, NativeFunction, Source,
-};
-use boa_gc::{Finalize, GcRefMut, Trace};
-use derive_more::{Deref, DerefMut};
-use http::Uri;
-use jstz_api::{
-    http::{
-        body::{Body, BodyWithType},
-        header::Headers,
-        request::{Request, RequestClass},
-        response::{Response, ResponseClass, ResponseOptions},
-    },
-    js_log::set_js_logger,
-};
-use jstz_core::{
-    host::HostRuntime, host_defined, kv::Transaction, native::JsNativeObject, runtime,
-    Module, Realm, Runtime,
-};
-use jstz_crypto::{hash::Hash, smart_function_hash::SmartFunctionHash};
-use tezos_smart_rollup::prelude::debug_msg;
-
-use crate::{
-    api::{self, TraceData},
-    context::account::{Account, Addressable, ParsedCode},
-    js_logger::JsonLogger,
-    operation::{OperationHash, RunFunction},
-    receipt::RunFunctionReceipt,
-    request_logger::{log_request_end, log_request_start},
-    Error, Result,
-};
-
-pub use deploy::deploy_smart_function as deploy;
-
-pub mod headers {
-
-    use crate::context::account::Addressable;
-
-    use super::*;
-    pub const REFERRER: &str = "Referer";
-
-    pub fn test_and_set_referrer(
-        request: &Request,
-        referrer: &impl Addressable,
-    ) -> JsResult<()> {
-        if request.headers().deref().contains_key(REFERRER) {
-            return Err(JsError::from_native(
-                JsNativeError::error().with_message("Referer already set"),
-            ));
-        }
-
-        request
-            .headers()
-            .deref_mut()
-            .set(REFERRER, &referrer.to_base58())
-    }
-}
-
-// Applies on_fullfilled or on_rejected based on either an error was raised or not.
-// If the value is a promise, then we apply the on_fulfilled and on_rejected to the promise.
-pub fn try_apply_to_value_or_promise<F1, F2>(
-    value_or_promise: JsResult<JsValue>,
-    on_fulfilled: F1,
-    on_rejected: F2,
-    context: &mut Context,
-) -> JsResult<JsValue>
-where
-    F1: Fn(JsValue, &mut Context) -> JsResult<JsValue> + 'static,
-    F2: Fn(&mut Context) -> JsResult<()> + 'static,
-{
-    match value_or_promise {
-        Ok(value) => match value.as_promise() {
-            Some(promise) => {
-                let result = promise.then(
-                    Some(
-                        FunctionObjectBuilder::new(context.realm(), unsafe {
-                            NativeFunction::from_closure(
-                                move |_, args, context| -> JsResult<JsValue> {
-                                    let value = args.get_or_undefined(0).clone();
-                                    on_fulfilled(value, context)
-                                },
-                            )
-                        })
-                        .build(),
-                    ),
-                    Some(
-                        FunctionObjectBuilder::new(context.realm(), unsafe {
-                            NativeFunction::from_closure(
-                                move |_, args, context| -> JsResult<JsValue> {
-                                    let value = args.get_or_undefined(0).clone();
-                                    let reason = JsError::from_opaque(value);
-                                    on_rejected(context)?;
-                                    Err(reason)
-                                },
-                            )
-                        })
-                        .build(),
-                    ),
-                    context,
-                );
-                Ok(result.into())
-            }
-            None => on_fulfilled(value, context),
-        },
-        Err(err) => {
-            on_rejected(context)?;
-            Err(err)
-        }
-    }
-}
-
-fn compute_seed(address: &SmartFunctionHash, operation_hash: &OperationHash) -> u64 {
-    let mut seed: u64 = 0;
-    for byte in operation_hash.as_array().iter().chain(address.as_bytes()) {
-        seed = seed.rotate_left(8).bitxor(*byte as u64)
-    }
-
-    seed
-}
-
-fn register_http_api(realm: &Realm, context: &mut Context) {
-    realm.register_api(jstz_api::http::HttpApi, context);
-}
-
-pub fn register_web_apis(realm: &Realm, context: &mut Context) {
-    realm.register_api(jstz_api::url::UrlApi, context);
-    realm.register_api(jstz_api::urlpattern::UrlPatternApi, context);
-    realm.register_api(jstz_api::http::HttpApi, context);
-    realm.register_api(jstz_api::encoding::EncodingApi, context);
-    realm.register_api(jstz_api::ConsoleApi, context);
-    realm.register_api(jstz_api::file::FileApi, context);
-}
-
-pub fn register_jstz_apis(
-    realm: &Realm,
-    address: &SmartFunctionHash,
-    seed: u64,
-    context: &mut Context,
-) {
-    realm.register_api(
-        api::KvApi {
-            address: address.clone(),
-        },
-        context,
-    );
-    realm.register_api(jstz_api::RandomApi { seed }, context);
-    realm.register_api(
-        api::LedgerApi {
-            address: address.clone(),
-        },
-        context,
-    );
-    realm.register_api(
-        api::SmartFunctionApi {
-            address: address.clone(),
-        },
-        context,
-    );
-}
-
-#[derive(Debug, PartialEq, Eq, Clone, Deref, DerefMut, Trace, Finalize)]
-pub struct Script(Module);
-
-impl Script {
-    fn get_default_export(&self, context: &mut Context) -> JsResult<JsValue> {
-        self.namespace(context).get(js_string!("default"), context)
-    }
-
-    fn invoke_handler(
-        &self,
-        this: &JsValue,
-        args: &[JsValue],
-        context: &mut Context,
-    ) -> JsResult<JsValue> {
-        let default_export = self.get_default_export(context)?;
-
-        let handler = default_export.as_object().ok_or_else(|| {
-            JsError::from_native(
-                JsNativeError::typ()
-                    .with_message("Failed to convert `default` export to js object"),
-            )
-        })?;
-
-        handler.call(this, args, context)
-    }
-
-    pub fn load(src: &ParsedCode, context: &mut Context) -> JsResult<Self> {
-        let module = Module::parse(
-            Source::from_bytes(src.as_bytes()),
-            Some(Realm::new(context)?),
-            context,
-        )?;
-        Ok(Self(module))
-    }
-
-    fn register_apis(
-        &self,
-        address: &SmartFunctionHash,
-        operation_hash: &OperationHash,
-        context: &mut Context,
-    ) {
-        register_web_apis(self.realm(), context);
-        register_jstz_apis(
-            self.realm(),
-            address,
-            compute_seed(address, operation_hash),
-            context,
-        );
-
-        let context = &mut self.realm().context_handle(context);
-
-        host_defined!(context, mut host_defined);
-        host_defined.insert(TraceData {
-            address: address.clone(),
-            operation_hash: operation_hash.clone(),
-        });
-    }
-
-    /// Initialize the script, registering all associated runtime APIs
-    /// and evaluating the module of the script
-    pub fn init(&self, context: &mut Context) -> JsPromise {
-        self.realm().eval_module(self, context)
-    }
-
-    /// Runs the script
-    pub fn run(&self, request: &JsValue, context: &mut Context) -> JsResult<JsValue> {
-        set_js_logger(&JsonLogger);
-        self.invoke_handler(&JsValue::undefined(), &[request.clone()], context)
-    }
-
-    /// Loads, initializes and runs the script
-    pub fn load_init_run(
-        src: &ParsedCode,
-        address: SmartFunctionHash,
-        operation_hash: OperationHash,
-        request: &JsValue,
-        context: &mut Context,
-    ) -> JsResult<JsValue> {
-        // 1. Load script
-        let script = Self::load(src, context)?;
-
-        // 2. Register the APIs for the script's realm
-        script.register_apis(&address, &operation_hash, context);
-
-        // 3. Evaluate the script's module
-        let script_promise = script.init(context);
-
-        // 4. Once evaluated, call the script's handler
-        let result = script_promise.then(
-            Some(
-                FunctionObjectBuilder::new(context.realm(), unsafe {
-                    NativeFunction::from_closure_with_captures(
-                        |_, _, (script, request), context| script.run(request, context),
-                        (script, request.clone()),
-                    )
-                })
-                .build(),
-            ),
-            None,
-            context,
-        );
-
-        Ok(result.into())
-    }
-}
-
-pub fn runtime_and_request_from_run_operation(
-    run_operation: RunFunction,
-) -> Result<(Runtime, JsNativeObject<Request>)> {
-    let RunFunction {
-        uri,
-        method,
-        headers,
-        body,
-        gas_limit,
-    } = run_operation;
-
-    let mut rt = Runtime::new(gas_limit)?;
-    register_http_api(&rt.realm().clone(), &mut rt);
-
-    let mut http_request_builder = http::Request::builder().uri(uri).method(method);
-
-    *http_request_builder
-        .headers_mut()
-        .ok_or(Error::InvalidHttpRequest)? = headers;
-
-    let http_request = http_request_builder
-        .body(body)
-        .map_err(|_| Error::InvalidHttpRequest)?;
-    let request = JsNativeObject::new::<RequestClass>(
-        Request::from_http_request(http_request, &mut rt)?,
-        &mut rt,
-    )?;
-
-    Ok((rt, request))
-}
-
-fn run_function_from_request(
-    request_deref: &mut GcRefMut<'_, ErasedObject, Request>,
-    gas_limit: usize,
-) -> JsResult<RunFunction> {
-    let method = request_deref.method().clone();
-    let uri = Uri::try_from(request_deref.url().clone().to_string()).map_err(|_| {
-        JsError::from_native(JsNativeError::error().with_message("Invalid host"))
-    })?;
-    let body = request_deref.body().clone().to_http_body();
-    let headers = request_deref.headers().deref_mut().to_http_headers();
-    Ok(RunFunction {
-        uri,
-        method,
-        body,
-        headers,
-        gas_limit,
-    })
-}
-
-fn response_from_run_receipt(
-    run_receipt: RunFunctionReceipt,
-    context: &mut Context,
-) -> JsResult<Response> {
-    let body = Body::from_http_body(run_receipt.body, context)?;
-    let options = ResponseOptions::from(run_receipt.status_code, run_receipt.headers);
-    Response::new(
-        BodyWithType {
-            body,
-            content_type: None,
-        },
-        options,
-        context,
-    )
-}
-
-pub const X_JSTZ_TRANSFER: &str = "X-JSTZ-TRANSFER";
-pub const X_JSTZ_AMOUNT: &str = "X-JSTZ-AMOUNT";
-
-pub struct HostScript;
-
-impl HostScript {
-    pub fn run(
-        source_address: &impl Addressable,
-        request: &mut GcRefMut<'_, ErasedObject, Request>,
-        context: &mut Context,
-    ) -> JsResult<JsValue> {
-        let run = run_function_from_request(request, 1)?;
-        let response = runtime::with_js_hrt_and_tx(|hrt, tx| -> JsResult<Response> {
-            // 1. Begin a new transaction
-            tx.begin();
-            // 2. Execute jstz host smart function
-            let result = jstz_run::execute_without_ticketer(hrt, tx, source_address, run);
-
-            // 3. Commit or rollback the transaction
-            match result {
-                Ok(run_receipt) => {
-                    if run_receipt.status_code.is_success() {
-                        tx.commit(hrt)?;
-                    } else {
-                        tx.rollback()?;
-                    }
-                    response_from_run_receipt(run_receipt, context)
-                }
-                Err(err) => {
-                    tx.rollback()?;
-                    Err(err.into())
-                }
-            }
-        })?;
-
-        let js_response = JsNativeObject::new::<ResponseClass>(response, context)?;
-        Ok(js_response.inner().clone())
-    }
-
-    /// Extracts the XTZ transfer amount from the request headers.
-    /// Returns None if the header is not present or Some(amount) if a valid amount is found.
-    pub fn extract_transfer_amount(headers: &Headers) -> JsResult<Option<NonZeroU64>> {
-        let header = headers.get(X_JSTZ_TRANSFER)?;
-
-        if header.headers.is_empty() {
-            return Ok(None);
-        }
-
-        if header.headers.len() > 1 {
-            return Err(JsError::from_native(JsNativeError::typ().with_message(
-                "Invalid transfer header: expected exactly one value",
-            )));
-        }
-
-        let amount = header.headers[0]
-            .parse::<NonZeroU64>()
-            .map(Some)
-            .map_err(|e| {
-                JsError::from_native(
-                    JsNativeError::typ()
-                        .with_message(format!("Invalid transfer amount: {}", e)),
-                )
-            })?;
-
-        Ok(amount)
-    }
-
-    fn verify_headers(headers: &Headers) -> JsResult<()> {
-        if headers.contains_key(X_JSTZ_AMOUNT) {
-            return Err(JsError::from_native(
-                JsNativeError::error()
-                    .with_message("X-JSTZ-AMOUNT header should not be present"),
-            ));
-        }
-        Ok(())
-    }
-
-    /// Transfer xtz from `src` to `dst` if the `X_JSTZ_TRANSFER` header is present & amount > 0
-    /// On success, `X_JSTZ_TRANSFER` is set to `X_JSTZ_AMOUNT`
-    /// Rejects if `X_JSTZ_AMOUNT` is already present in the headers or transfer failed
-    pub fn handle_transfer(
-        headers: &mut Headers,
-        src: &impl Addressable,
-        dst: &impl Addressable,
-    ) -> JsResult<Option<NonZeroU64>> {
-        Self::verify_headers(headers)?;
-        let amt = match Self::extract_transfer_amount(headers)? {
-            Some(a) => a,
-            None => return Ok(None),
-        };
-        runtime::with_js_hrt_and_tx(|hrt, tx| {
-            Account::transfer(hrt, tx, src, dst, amt.into())
-                .and_then(|_| {
-                    headers.remove(X_JSTZ_TRANSFER)?;
-                    headers.append(X_JSTZ_AMOUNT, &amt.to_string())?;
-                    Ok(())
-                })
-                .map_err(|e| {
-                    JsError::from_native(
-                        JsNativeError::eval()
-                            .with_message(format!("Transfer failed: {}", e)),
-                    )
-                })
-        })?;
-        Ok(Some(amt))
-    }
-}
-
-pub mod run {
-
-    use jstz_api::http::response::ResponseBuilder;
-
-    use super::*;
-    use crate::{
-        context::account::Address,
-        executor::JSTZ_HOST,
-        operation::{self, OperationHash},
-        receipt::RunFunctionReceipt,
-    };
-
-    /// skips the execution of the smart function for noop requests
-    pub const NOOP_PATH: &str = "/-/noop";
-
-    /// Handles the transfer (or a refund) (if any) and returns [`Ok(None)`] if the transfer
-    /// succeeded. If the transfer fails, it rolls back the transaction and returns an error response
-    /// as [`Ok(Some(err_response))`].
-    fn handle_transfer_or_rollback_and_return_response(
-        headers: &mut Headers,
-        source_address: &impl Addressable,
-        dest_address: &impl Addressable,
-        context: &mut Context,
-    ) -> Result<Option<JsValue>> {
-        if HostScript::handle_transfer(headers, source_address, dest_address).is_err() {
-            // If the transfer fails, return an error response
-            runtime::with_js_tx(|tx| tx.rollback())?;
-            Ok(Some(
-                JsNativeObject::new::<ResponseClass>(
-                    ResponseBuilder::error(context)?,
-                    context,
-                )?
-                .into(),
-            ))
-        } else {
-            Ok(None)
-        }
-    }
-
-    // Handles a fetch request to a smart function or a user address in the Jstz protocol.
-    pub fn fetch(
-        source_address: &(impl Addressable + 'static),
-        operation_hash: OperationHash,
-        request: &JsNativeObject<Request>,
-        context: &mut Context,
-    ) -> JsResult<JsValue> {
-        let mut request_deref = request.deref_mut();
-        if request_deref.url().scheme() != "jstz" {
-            return Err(Error::InvalidScheme.into());
-        }
-        match request_deref.url().domain() {
-            Some(JSTZ_HOST) => {
-                HostScript::run(source_address, &mut request_deref, context)
-            }
-            Some(dest_address) => {
-                let dest_address = Address::from_base58(dest_address).map_err(|_| {
-                    JsError::from_native(
-                        JsNativeError::error().with_message("Invalid host"),
-                    )
-                })?;
-
-                runtime::with_js_tx(|tx| tx.begin());
-
-                // 1. Handle the transfer operation in request headers
-                if let Some(error) = handle_transfer_or_rollback_and_return_response(
-                    &mut request_deref.headers().deref_mut(),
-                    source_address,
-                    &dest_address,
-                    context,
-                )? {
-                    // If the transfer fails, return an error response
-                    return Ok(error);
-                }
-
-                // 3. Handle request
-                let is_noop = request_deref.url().path() == NOOP_PATH;
-                match dest_address {
-                    Address::SmartFunction(dest_address) if !is_noop => {
-                        // 5. Set the referrer of the request to the current smart function address
-                        headers::test_and_set_referrer(&request_deref, source_address)?;
-
-                        // 6. Load, init and run the smart function
-                        let src_code = runtime::with_js_hrt_and_tx(
-                            |hrt, tx| -> Result<ParsedCode> {
-                                Ok(Account::function_code(hrt, tx, &dest_address)?
-                                    .clone())
-                            },
-                        )?;
-
-                        log_request_start(
-                            dest_address.clone(),
-                            operation_hash.to_string(),
-                        );
-                        let response = Script::load_init_run(
-                            &src_code,
-                            dest_address.clone(),
-                            operation_hash.clone(),
-                            request.inner(),
-                            context,
-                        );
-
-                        // TODO: avoid cloning
-                        // https://linear.app/tezos/issue/JSTZ-331/avoid-cloning-for-address-in-proto
-                        let source_address_clone = source_address.clone();
-                        let dest_address_clone = dest_address.clone();
-                        let response = try_apply_to_value_or_promise(
-                            response,
-                            move |value, context| -> JsResult<JsValue> {
-                                match Response::try_from_js(&value) {
-                                    Ok(response) => {
-                                        if let Some(error) =
-                                            handle_transfer_or_rollback_and_return_response(
-                                                &mut response.headers().deref_mut(),
-                                                &dest_address_clone,
-                                                &source_address_clone,
-                                                context,
-                                            )?
-                                        {
-                                            return Ok(error);
-                                        }
-
-                                        // If the smart function returns a valid response, commit the inner
-                                        // transaction iff the response is ok
-                                        runtime::with_js_hrt_and_tx(|hrt, tx| {
-                                            if response.ok() {
-                                                tx.commit(hrt)
-                                            } else {
-                                                tx.rollback()
-                                            }
-                                        })?;
-                                    }
-                                    _ => {
-                                        // If the smart function doesn't return a valid response,
-                                        // rollback the inner transaction (abort)
-                                        runtime::with_js_tx(|tx| tx.rollback())?;
-                                    }
-                                }
-                                Ok(value)
-                            },
-                            |_context| Ok(runtime::with_js_tx(|tx| tx.rollback())?),
-                            context,
-                        );
-
-                        log_request_end(dest_address.clone(), operation_hash.to_string());
-                        response
-                    }
-                    _ => {
-                        // Request is a noop request or to a user address
-                        runtime::with_js_hrt_and_tx(|hrt, tx| tx.commit(hrt))?;
-
-                        // Return a default response
-                        let response = Response::new(
-                            Default::default(),
-                            Default::default(),
-                            context,
-                        )?;
-                        JsNativeObject::new::<ResponseClass>(response, context)
-                            .map(|obj| obj.inner().clone())
-                    }
-                }
-            }
-            None => Err(JsError::from_native(
-                JsNativeError::error().with_message("Invalid host"),
-            ))?,
-        }
-    }
-
-    pub fn run_toplevel_fetch(
-        hrt: &mut impl HostRuntime,
-        tx: &mut Transaction,
-        source_address: &(impl Addressable + 'static),
-        run_operation: RunFunction,
-        operation_hash: OperationHash,
-    ) -> Result<RunFunctionReceipt> {
-        let gas_limit = run_operation.gas_limit;
-        let (mut rt, request) = runtime_and_request_from_run_operation(run_operation)?;
-
-        let result = {
-            let rt = &mut rt;
-            runtime::enter_js_host_context(hrt, tx, || {
-                jstz_core::future::block_on(async move {
-                    let result = fetch(source_address, operation_hash, &request, rt)?;
-                    rt.resolve_value(&result).await
-                })
-            })
-        }
-        .map_err(|err| {
-            if rt.instructions_remaining() == 0 {
-                Error::GasLimitExceeded
-            } else {
-                err.into()
-            }
-        })?;
-
-        debug_msg!(hrt, "🚀 Smart function executed successfully with value: {:?} (in {:?} instructions)\n", result, gas_limit - rt.instructions_remaining());
-
-        let response = Response::try_from_js(&result)?;
-        let (http_parts, body) = Response::to_http_response(&response).into_parts();
-        Ok(RunFunctionReceipt {
-            body,
-            status_code: http_parts.status,
-            headers: http_parts.headers,
-        })
-    }
-
-    pub fn execute(
-        hrt: &mut impl HostRuntime,
-        tx: &mut Transaction,
-        source: &(impl Addressable + 'static),
-        run_operation: operation::RunFunction,
-        operation_hash: OperationHash,
-    ) -> Result<RunFunctionReceipt> {
-        run_toplevel_fetch(hrt, tx, source, run_operation, operation_hash)
-    }
-
-    #[cfg(test)]
-    mod test {
-        use super::*;
-        use http::{HeaderMap, Method};
-        use jstz_core::kv::Transaction;
-        use jstz_crypto::hash::Blake2b;
-        use jstz_mock::host::JstzMockHost;
-
-        use crate::{
-            context::account::{Account, Address, ParsedCode},
-            executor::smart_function,
-            operation::RunFunction,
-        };
-
-        #[test]
-        fn transfer_xtz_to_and_from_smart_function_succeeds() {
-            let source = Address::User(jstz_mock::account1());
-            // 1. Deploy the smart function
-            let mut jstz_mock_host = JstzMockHost::default();
-            let host = jstz_mock_host.rt();
-            let mut tx = Transaction::default();
-            let transfer_amount = 3;
-            let refund_amount = 2;
-            tx.begin();
-            Account::add_balance(host, &mut tx, &source, transfer_amount)
-                .expect("add balance");
-            let source_balance = Account::balance(host, &mut tx, &source).unwrap();
-            assert_eq!(source_balance, transfer_amount);
-            tx.commit(host).unwrap();
-
-            // 1. Deploy the smart function that transfers the balance to the source
-            let code = format!(
-                r#"
-            const handler = async (request) => {{
-                const transferred_amount = request.headers.get("X-JSTZ-AMOUNT");
-                if (transferred_amount !== "{transfer_amount}") {{
-                    return Response.error("Invalid transferred amount");
-                }}
-                const headers = {{"X-JSTZ-TRANSFER": "{refund_amount}"}};
-                return new Response(null, {{headers}});
-            }};
-            export default handler;
-            "#
-            );
-            let parsed_code = ParsedCode::try_from(code.to_string()).unwrap();
-            tx.begin();
-            let smart_function =
-                smart_function::deploy(host, &mut tx, &source, parsed_code, 0).unwrap();
-
-            let balance_before =
-                Account::balance(host, &mut tx, &smart_function).unwrap();
-            assert_eq!(balance_before, 0);
-
-            tx.commit(host).unwrap();
-
-            // 2. Call the smart function
-            tx.begin();
-            let mut headers = HeaderMap::new();
-            headers.insert(
-                X_JSTZ_TRANSFER,
-                transfer_amount.to_string().try_into().unwrap(),
-            );
-            let run_function = RunFunction {
-                uri: format!("jstz://{}/", &smart_function).try_into().unwrap(),
-                method: Method::GET,
-                headers,
-                body: None,
-                gas_limit: 1000,
-            };
-            let fake_op_hash = Blake2b::from(b"fake_op_hash".as_ref());
-            let response = execute(
-                host,
-                &mut tx,
-                &source,
-                run_function.clone(),
-                fake_op_hash.clone(),
-            )
-            .expect("run function expected");
-
-            assert!(response.headers.get(X_JSTZ_TRANSFER).is_none());
-            assert!(response.headers.get(X_JSTZ_AMOUNT).is_some_and(|amt| amt
-                .to_str()
-                .unwrap()
-                .parse::<u64>()
-                .unwrap()
-                == refund_amount));
-            tx.commit(host).unwrap();
-
-            // 3. assert the transfer to the sf and refund to the source
-            tx.begin();
-            let balance_after = Account::balance(host, &mut tx, &smart_function).unwrap();
-            assert_eq!(
-                balance_after - balance_before,
-                transfer_amount - refund_amount
-            );
-            assert_eq!(
-                Account::balance(host, &mut tx, &source).unwrap(),
-                refund_amount
-            );
-
-            // 4. transferring to the smart function should fail (source has insufficient funds)
-            let result = execute(
-                host,
-                &mut tx,
-                &source,
-                run_function.clone(),
-                fake_op_hash.clone(),
-            )
-            .unwrap();
-            assert!(result.status_code.is_server_error());
-
-            // 5. transferring from the smart function should fail with insufficient funds and the balance is rolled back
-            let balance_before = Account::balance(host, &mut tx, &source).unwrap();
-            // drain the balance of the smart function
-            Account::set_balance(host, &mut tx, &smart_function, 0).unwrap();
-            let mut headers = HeaderMap::new();
-            headers.insert(
-                X_JSTZ_TRANSFER,
-                transfer_amount.to_string().try_into().unwrap(),
-            );
-            let result = execute(
-                host,
-                &mut tx,
-                &source,
-                RunFunction {
-                    headers,
-                    ..run_function
-                },
-                fake_op_hash.clone(),
-            )
-            .unwrap();
-            assert!(result.status_code.is_server_error());
-
-            // tx rolled back as smart function has insufficient funds
-            let balance_after = Account::balance(host, &mut tx, &source).unwrap();
-            assert_eq!(balance_after, balance_before);
-        }
-
-        #[test]
-        fn transfer_xtz_to_smart_function_succeeds_with_noop_path() {
-            let source = Address::User(jstz_mock::account1());
-            // 1. Deploy the smart function
-            let mut jstz_mock_host = JstzMockHost::default();
-            let host = jstz_mock_host.rt();
-            let mut tx = Transaction::default();
-            let initial_balance = 1;
-            tx.begin();
-            Account::add_balance(host, &mut tx, &source, initial_balance)
-                .expect("add balance");
-            let source_balance = Account::balance(host, &mut tx, &source).unwrap();
-            assert_eq!(source_balance, initial_balance);
-            tx.commit(host).unwrap();
-
-            // 1. Deploy the smart function that refunds the balance to the source
-            let code = format!(
-                r#"
-                const handler = async () => {{
-                    await fetch(new Request("jstz://{source}", {{
-                        headers: {{"X-JSTZ-TRANSFER": "{initial_balance}"}}
-                    }}));
-                    return new Response();
-                }};
-                export default handler;
-                "#
-            );
-            let parsed_code = ParsedCode::try_from(code.to_string()).unwrap();
-            tx.begin();
-            let smart_function =
-                smart_function::deploy(host, &mut tx, &source, parsed_code, 0).unwrap();
-
-            let balance_before =
-                Account::balance(host, &mut tx, &smart_function).unwrap();
-            assert_eq!(balance_before, 0);
-
-            tx.commit(host).unwrap();
-
-            // transfer should happen with `/-/noop` path
-            tx.begin();
-            let mut headers = HeaderMap::new();
-            headers.insert(
-                X_JSTZ_TRANSFER,
-                initial_balance.to_string().try_into().unwrap(),
-            );
-            let run_function = RunFunction {
-                uri: format!("jstz://{}/-/noop", &smart_function)
-                    .try_into()
-                    .unwrap(),
-                method: Method::GET,
-                headers,
-                body: None,
-                gas_limit: 1000,
-            };
-            let fake_op_hash = Blake2b::from(b"fake_op_hash".as_ref());
-            execute(host, &mut tx, &source, run_function.clone(), fake_op_hash)
-                .expect("run function expected");
-            tx.commit(host).unwrap();
-
-            tx.begin();
-            let balance_after = Account::balance(host, &mut tx, &smart_function).unwrap();
-            assert_eq!(balance_after - balance_before, initial_balance);
-            assert_eq!(Account::balance(host, &mut tx, &source).unwrap(), 0);
-        }
-
-        #[test]
-        fn transfer_xtz_to_user_succeeds() {
-            let source = Address::User(jstz_mock::account1());
-            let destination = Address::User(jstz_mock::account2());
-            // 1. Deploy the smart function
-            let mut jstz_mock_host = JstzMockHost::default();
-            let host = jstz_mock_host.rt();
-            let mut tx = Transaction::default();
-            let initial_balance = 1;
-            tx.begin();
-            Account::add_balance(host, &mut tx, &source, initial_balance)
-                .expect("add balance");
-            let source_balance = Account::balance(host, &mut tx, &source).unwrap();
-            assert_eq!(source_balance, initial_balance);
-            tx.commit(host).unwrap();
-
-            // 2. sending request to transfer from source to the destination
-            tx.begin();
-            let mut headers = HeaderMap::new();
-            headers.insert(
-                X_JSTZ_TRANSFER,
-                initial_balance.to_string().try_into().unwrap(),
-            );
-            let run_function = RunFunction {
-                uri: format!("jstz://{}/", &destination).try_into().unwrap(),
-                method: Method::GET,
-                headers,
-                body: None,
-                gas_limit: 1000,
-            };
-            let fake_op_hash = Blake2b::from(b"fake_op_hash".as_ref());
-            let result =
-                execute(host, &mut tx, &source, run_function.clone(), fake_op_hash);
-            assert!(result.is_ok());
-
-            tx.commit(host).unwrap();
-
-            tx.begin();
-            let balance_after = Account::balance(host, &mut tx, &source).unwrap();
-            assert_eq!(balance_after, 0);
-            assert_eq!(
-                Account::balance(host, &mut tx, &destination).unwrap(),
-                initial_balance
-            );
-
-            // 3. transferring again should fail
-            let fake_op_hash2 = Blake2b::from(b"fake_op_hash2".as_ref());
-            let result =
-                execute(host, &mut tx, &source, run_function, fake_op_hash2).unwrap();
-            assert!(result.status_code.is_server_error());
-        }
-
-        #[test]
-        fn invalid_request_should_fails() {
-            let source = Address::User(jstz_mock::account1());
-            // 1. Deploy the smart function
-            let mut jstz_mock_host = JstzMockHost::default();
-            let host = jstz_mock_host.rt();
-            let mut tx = Transaction::default();
-            let initial_balance = 1;
-            tx.begin();
-            Account::add_balance(host, &mut tx, &source, initial_balance)
-                .expect("add balance");
-            tx.commit(host).unwrap();
-
-            let code = r#"
-                const handler = () => {{
-                    return new Response();
-                }};
-                export default handler;
-                "#;
-
-            // 1. Deploy smart function
-            let parsed_code = ParsedCode::try_from(code.to_string()).unwrap();
-            tx.begin();
-            let smart_function =
-                smart_function::deploy(host, &mut tx, &source, parsed_code, 0).unwrap();
-
-            tx.commit(host).unwrap();
-
-            // Calling the smart function should error or return an error response
-            tx.begin();
-
-            let sf_balance_before =
-                Account::balance(host, &mut tx, &smart_function).unwrap();
-            let source_balance_before = Account::balance(host, &mut tx, &source).unwrap();
-            let mut invalid_headers = HeaderMap::new();
-            invalid_headers.insert(
-                X_JSTZ_AMOUNT,
-                initial_balance.to_string().try_into().unwrap(),
-            );
-            let run_function = RunFunction {
-                uri: format!("jstz://{}/", &smart_function).try_into().unwrap(),
-                method: Method::GET,
-                headers: invalid_headers,
-                body: None,
-                gas_limit: 1000,
-            };
-            let result = execute(
-                host,
-                &mut tx,
-                &source,
-                run_function.clone(),
-                Blake2b::from(b"fake_op_hash".as_ref()),
-            );
-            let sf_balance_after =
-                Account::balance(host, &mut tx, &smart_function).unwrap();
-            let source_balance_after = Account::balance(host, &mut tx, &source).unwrap();
-
-            assert_eq!(sf_balance_before, sf_balance_after);
-            assert_eq!(source_balance_before, source_balance_after);
-            let call_failed = match result {
-                Ok(receipt) => receipt.status_code.is_server_error(),
-                _ => true,
-            };
-            assert!(call_failed);
-        }
-
-        #[test]
-        fn invalid_response_should_fails() {
-            let source = Address::User(jstz_mock::account1());
-            // 1. Deploy the smart function
-            let mut jstz_mock_host = JstzMockHost::default();
-            let host = jstz_mock_host.rt();
-            let mut tx = Transaction::default();
-            let initial_balance = 1;
-            tx.begin();
-            Account::add_balance(host, &mut tx, &source, initial_balance)
-                .expect("add balance");
-            tx.commit(host).unwrap();
-
-            let code = format!(
-                r#"
-                const handler = () => {{
-                    const headers = new Headers();
-                    return new Response(null, {{
-                        headers: {{ "X-JSTZ-AMOUNT": "{initial_balance}" }},
-                    }});
-                }};
-                export default handler;
-                "#
-            );
-
-            // 1. Deploy smart function
-            let parsed_code = ParsedCode::try_from(code.to_string()).unwrap();
-            tx.begin();
-            let smart_function = smart_function::deploy(
-                host,
-                &mut tx,
-                &source,
-                parsed_code,
-                initial_balance,
-            )
-            .unwrap();
-
-            let sf_balance_before =
-                Account::balance(host, &mut tx, &smart_function).unwrap();
-            let source_balance_before = Account::balance(host, &mut tx, &source).unwrap();
-
-            tx.commit(host).unwrap();
-
-            // Calling the smart function should error or return an error response
-            tx.begin();
-            let run_function = RunFunction {
-                uri: format!("jstz://{}/", &smart_function).try_into().unwrap(),
-                method: Method::GET,
-                headers: Default::default(),
-                body: None,
-                gas_limit: 1000,
-            };
-            let result = execute(
-                host,
-                &mut tx,
-                &source,
-                run_function.clone(),
-                Blake2b::from(b"fake_op_hash".as_ref()),
-            );
-            let sf_balance_after =
-                Account::balance(host, &mut tx, &smart_function).unwrap();
-            let source_balance_after = Account::balance(host, &mut tx, &source).unwrap();
-
-            assert_eq!(sf_balance_before, sf_balance_after);
-            assert_eq!(source_balance_before, source_balance_after);
-            let call_failed = match result {
-                Ok(receipt) => receipt.status_code.is_server_error(),
-                _ => true,
-            };
-            assert!(call_failed);
-        }
-
-        #[test]
-        fn transfer_xtz_and_smart_function_call_is_atomic1() {
-            let invalid_code = r#"
-           const handler = () => {{
-               invalid();
-           }};
-           export default handler;
-           "#;
-            transfer_xtz_and_run_erroneous_sf(invalid_code);
-        }
-
-        #[test]
-        fn transfer_xtz_and_smart_function_call_is_atomic2() {
-            let invalid_code = r#"
-           const handler = () => {{
-                return Response.error("error!");
-           }};
-           export default handler;
-           "#;
-            transfer_xtz_and_run_erroneous_sf(invalid_code);
-        }
-
-        #[test]
-        fn transfer_xtz_and_smart_function_call_is_atomic3() {
-            let invalid_code = r#"
-           const handler = () => {{
-            return 3;
-           }};
-           export default handler;
-           "#;
-            transfer_xtz_and_run_erroneous_sf(invalid_code);
-        }
-
-        fn transfer_xtz_and_run_erroneous_sf(code: &str) {
-            let source = Address::User(jstz_mock::account1());
-            // 1. Deploy the smart function
-            let mut jstz_mock_host = JstzMockHost::default();
-            let host = jstz_mock_host.rt();
-            let mut tx = Transaction::default();
-            let initial_balance = 1;
-            tx.begin();
-            Account::add_balance(host, &mut tx, &source, initial_balance)
-                .expect("add balance");
-            tx.commit(host).unwrap();
-
-            // 1. Deploy smart function
-            let parsed_code = ParsedCode::try_from(code.to_string()).unwrap();
-            tx.begin();
-            let smart_function =
-                smart_function::deploy(host, &mut tx, &source, parsed_code, 0).unwrap();
-
-            tx.commit(host).unwrap();
-
-            // Calling the smart function should error or return an error response
-            tx.begin();
-            let mut headers = HeaderMap::new();
-            headers.insert(
-                X_JSTZ_TRANSFER,
-                initial_balance.to_string().try_into().unwrap(),
-            );
-            let run_function = RunFunction {
-                uri: format!("jstz://{}/", &smart_function).try_into().unwrap(),
-                method: Method::GET,
-                headers,
-                body: None,
-                gas_limit: 1000,
-            };
-            let result = execute(
-                host,
-                &mut tx,
-                &source,
-                run_function.clone(),
-                Blake2b::from(b"fake_op_hash".as_ref()),
-            );
-            let call_failed = match result {
-                Ok(receipt) => receipt.status_code.is_server_error(),
-                _ => true,
-            };
-            assert!(call_failed);
-
-            // The balance should not be affected
-            assert_eq!(
-                Account::balance(host, &mut tx, &source).unwrap(),
-                initial_balance
-            );
-            let balance_after = Account::balance(host, &mut tx, &smart_function).unwrap();
-            assert_eq!(balance_after, 0);
-        }
-    }
-}
-
-pub mod jstz_run {
-    use jstz_core::kv::Storage;
-    use serde::Deserialize;
-    use tezos_crypto_rs::hash::ContractKt1Hash;
-    use tezos_smart_rollup::storage::path::{OwnedPath, RefPath};
-
-    use super::*;
-    use crate::{
-        executor::{fa_withdraw::FaWithdraw, withdraw::Withdrawal, JSTZ_HOST},
-        operation::RunFunction,
-        receipt,
-    };
-
-    const WITHDRAW_PATH: &str = "/withdraw";
-    const FA_WITHDRAW_PATH: &str = "/fa-withdraw";
-
-    fn validate_withdraw_request<'de, T>(run: &'de RunFunction) -> Result<T>
-    where
-        T: Deserialize<'de>,
-    {
-        let method = run
-            .method
-            .as_str()
-            .parse::<http::Method>()
-            .map_err(|_| Error::InvalidHttpRequestMethod)?;
-
-        if method != http::Method::POST {
-            return Err(Error::InvalidHttpRequestMethod);
-        }
-
-        if run.body.is_none() {
-            return Err(Error::InvalidHttpRequestBody);
-        }
-        let withdrawal = serde_json::from_slice(run.body.as_ref().unwrap())
-            .map_err(|_| Error::InvalidHttpRequestBody)?;
-        Ok(withdrawal)
-    }
-
-    pub fn execute(
-        hrt: &mut impl HostRuntime,
-        tx: &mut Transaction,
-        ticketer: &ContractKt1Hash,
-        source: &impl Addressable,
-        run: RunFunction,
-    ) -> Result<receipt::RunFunctionReceipt> {
-        let uri = run.uri.clone();
-        if uri.host() != Some(JSTZ_HOST) {
-            return Err(Error::InvalidHost);
-        }
-        match uri.path() {
-            WITHDRAW_PATH => {
-                // TODO: https://linear.app/tezos/issue/JSTZ-77/check-gas-limit-when-performing-native-withdraws
-                // Check gas limit
-
-                let withdrawal = validate_withdraw_request::<Withdrawal>(&run)?;
-                crate::executor::withdraw::execute_withdraw(
-                    hrt, tx, source, withdrawal, ticketer,
-                )?;
-                let receipt = receipt::RunFunctionReceipt {
-                    body: None,
-                    status_code: http::StatusCode::OK,
-                    headers: http::HeaderMap::new(),
-                };
-                Ok(receipt)
-            }
-            FA_WITHDRAW_PATH => {
-                let fa_withdraw = validate_withdraw_request::<FaWithdraw>(&run)?;
-                let fa_withdraw_receipt_content = fa_withdraw.execute(
-                    hrt, tx, source, 1000, // fake gas limit
-                )?;
-                let receipt = receipt::RunFunctionReceipt {
-                    body: fa_withdraw_receipt_content.to_http_body(),
-                    status_code: http::StatusCode::OK,
-                    headers: http::HeaderMap::new(),
-                };
-                Ok(receipt)
-            }
-            _ => Err(Error::UnsupportedPath),
-        }
-    }
-
-    pub fn execute_without_ticketer(
-        hrt: &mut impl HostRuntime,
-        tx: &mut Transaction,
-        source: &impl Addressable,
-        run: RunFunction,
-    ) -> Result<receipt::RunFunctionReceipt> {
-        let ticketer_path = OwnedPath::from(&RefPath::assert_from(b"/ticketer"));
-        let ticketer: SmartFunctionHash =
-            Storage::get(hrt, &ticketer_path)?.expect("ticketer should be set");
-        execute(hrt, tx, &ticketer, source, run)
-    }
-
-    #[cfg(test)]
-    mod test {
-        use http::{header, HeaderMap, Method, Uri};
-        use jstz_core::kv::Transaction;
-        use jstz_crypto::hash::Hash;
-        use jstz_mock::host::JstzMockHost;
-        use serde_json::json;
-        use tezos_crypto_rs::hash::ContractKt1Hash;
-        use tezos_smart_rollup_mock::MockHost;
-
-        use crate::{
-            context::{account::Address, ticket_table::TicketTable},
-            executor::{
-                fa_withdraw::{FaWithdraw, RoutingInfo, TicketInfo},
-                smart_function::jstz_run::{execute_without_ticketer, Account},
-            },
-            operation::RunFunction,
-            Error,
-        };
-
-        use super::execute;
-
-        fn withdraw_request() -> RunFunction {
-            RunFunction {
-                uri: Uri::try_from("jstz://jstz/withdraw").unwrap(),
-                method: Method::POST,
-                headers: HeaderMap::from_iter([(
-                    header::CONTENT_TYPE,
-                    "application/json".try_into().unwrap(),
-                )]),
-                body: Some(
-                    json!({
-                        "amount": 10,
-                        "receiver": jstz_mock::account2().to_base58().to_string(),
-                    })
-                    .to_string()
-                    .as_bytes()
-                    .to_vec(),
-                ),
-                gas_limit: 10,
-            }
-        }
-
-        fn fa_withdraw_request() -> RunFunction {
-            let ticket_info = TicketInfo {
-                id: 1234,
-                content: Some(b"random ticket content".to_vec()),
-                ticketer: jstz_mock::kt1_account1().into(),
-            };
-            let routing_info = RoutingInfo {
-                receiver: Address::User(jstz_mock::account2()),
-                proxy_l1_contract: jstz_mock::kt1_account1().into(),
-            };
-            let fa_withdrawal = FaWithdraw {
-                amount: 10,
-                routing_info,
-                ticket_info,
-            };
-
-            RunFunction {
-                uri: Uri::try_from("jstz://jstz/fa-withdraw").unwrap(),
-                method: Method::POST,
-                headers: HeaderMap::from_iter([(
-                    header::CONTENT_TYPE,
-                    "application/json".try_into().unwrap(),
-                )]),
-                body: Some(json!(fa_withdrawal).to_string().as_bytes().to_vec()),
-                gas_limit: 10,
-            }
-        }
-
-        #[test]
-        fn execute_fails_on_invalid_host() {
-            let mut host = MockHost::default();
-            let mut tx = Transaction::default();
-            let source = Address::User(jstz_mock::account1());
-            let req = RunFunction {
-                uri: Uri::try_from("jstz://example.com/withdraw").unwrap(),
-                ..withdraw_request()
-            };
-            let ticketer =
-                ContractKt1Hash::from_base58_check(jstz_mock::host::NATIVE_TICKETER)
-                    .unwrap();
-            let result = execute(&mut host, &mut tx, &ticketer, &source, req);
-            assert!(matches!(result, Err(super::Error::InvalidHost)));
-        }
-
-        #[test]
-        fn execute_fails_on_unsupported_path() {
-            let mut host = MockHost::default();
-            let mut tx = Transaction::default();
-            let source = Address::User(jstz_mock::account1());
-            let req = RunFunction {
-                uri: Uri::try_from("jstz://jstz/blahblah").unwrap(),
-                ..withdraw_request()
-            };
-            let ticketer =
-                ContractKt1Hash::from_base58_check(jstz_mock::host::NATIVE_TICKETER)
-                    .unwrap();
-            let result = execute(&mut host, &mut tx, &ticketer, &source, req);
-            assert!(matches!(result, Err(super::Error::UnsupportedPath)));
-        }
-
-        #[test]
-        fn execute_wthdraw_fails_on_invalid_request_method() {
-            let mut host = MockHost::default();
-            let mut tx = Transaction::default();
-            let source = Address::User(jstz_mock::account1());
-            let req = RunFunction {
-                method: Method::GET,
-                ..withdraw_request()
-            };
-            let ticketer =
-                ContractKt1Hash::from_base58_check(jstz_mock::host::NATIVE_TICKETER)
-                    .unwrap();
-            let result = execute(&mut host, &mut tx, &ticketer, &source, req);
-            assert!(matches!(
-                result,
-                Err(super::Error::InvalidHttpRequestMethod)
-            ));
-        }
-
-        #[test]
-        fn execute_wthdraw_fails_on_invalid_request_body() {
-            let mut host = MockHost::default();
-            let mut tx = Transaction::default();
-            let source = Address::User(jstz_mock::account1());
-            let req = RunFunction {
-                body: Some(
-                    json!({
-                        "amount": 10,
-                        "not_receiver": jstz_mock::account2().to_base58()
-                    })
-                    .to_string()
-                    .as_bytes()
-                    .to_vec(),
-                ),
-                ..withdraw_request()
-            };
-            let ticketer =
-                ContractKt1Hash::from_base58_check(jstz_mock::host::NATIVE_TICKETER)
-                    .unwrap();
-            let result = execute(&mut host, &mut tx, &ticketer, &source, req);
-            assert!(matches!(result, Err(Error::InvalidHttpRequestBody)));
-
-            let req = RunFunction {
-                body: None,
-                ..withdraw_request()
-            };
-            let result = execute(&mut host, &mut tx, &ticketer, &source, req);
-            assert!(matches!(result, Err(Error::InvalidHttpRequestBody)));
-        }
-
-        #[test]
-        fn execute_withdraw_succeeds() {
-            let mut host = MockHost::default();
-            let mut tx = Transaction::default();
-            let source = Address::User(jstz_mock::account1());
-
-            tx.begin();
-            Account::add_balance(&host, &mut tx, &source, 10).unwrap();
-            tx.commit(&mut host).unwrap();
-
-            let req = withdraw_request();
-            let ticketer =
-                ContractKt1Hash::from_base58_check(jstz_mock::host::NATIVE_TICKETER)
-                    .unwrap();
-
-            execute(&mut host, &mut tx, &ticketer, &source, req)
-                .expect("Withdraw should not fail");
-
-            tx.begin();
-            assert_eq!(0, Account::balance(&host, &mut tx, &source).unwrap());
-
-            let level = host.run_level(|_| {});
-            assert_eq!(1, host.outbox_at(level).len());
-        }
-
-        #[test]
-        fn execute_without_ticketer_succeeds() {
-            let mut host = JstzMockHost::default();
-            let mut tx = Transaction::default();
-            let source = Address::User(jstz_mock::account1());
-            let rt = host.rt();
-
-            tx.begin();
-            Account::add_balance(rt, &mut tx, &source, 10).unwrap();
-            tx.commit(rt).unwrap();
-
-            let req = withdraw_request();
-
-            execute_without_ticketer(rt, &mut tx, &source, req)
-                .expect("Withdraw should not fail");
-
-            tx.begin();
-            assert_eq!(0, Account::balance(rt, &mut tx, &source).unwrap());
-
-            let level = rt.run_level(|_| {});
-            assert_eq!(1, rt.outbox_at(level).len());
-        }
-
-        #[test]
-        fn execute_fa_withdraw_fails_on_invalid_request_method() {
-            let mut host = MockHost::default();
-            let mut tx = Transaction::default();
-            let source = Address::User(jstz_mock::account1());
-            let req = RunFunction {
-                method: Method::GET,
-                ..fa_withdraw_request()
-            };
-            let ticketer =
-                ContractKt1Hash::from_base58_check(jstz_mock::host::NATIVE_TICKETER)
-                    .unwrap();
-            let result = execute(&mut host, &mut tx, &ticketer, &source, req);
-            assert!(matches!(
-                result,
-                Err(super::Error::InvalidHttpRequestMethod)
-            ));
-        }
-
-        #[test]
-        fn execute_fa_withdraw_fails_on_invalid_request_body() {
-            let mut host = MockHost::default();
-            let mut tx = Transaction::default();
-            let source = Address::User(jstz_mock::account1());
-            let req = RunFunction {
-                body: Some(
-                    json!({
-                        "amount": 10,
-                        "not_receiver": jstz_mock::account2().to_base58()
-                    })
-                    .to_string()
-                    .as_bytes()
-                    .to_vec(),
-                ),
-                ..fa_withdraw_request()
-            };
-            let ticketer =
-                ContractKt1Hash::from_base58_check(jstz_mock::host::NATIVE_TICKETER)
-                    .unwrap();
-            let result = execute(&mut host, &mut tx, &ticketer, &source, req);
-            assert!(matches!(result, Err(Error::InvalidHttpRequestBody)));
-
-            let req = RunFunction {
-                body: None,
-                ..withdraw_request()
-            };
-            let result = execute(&mut host, &mut tx, &ticketer, &source, req);
-            assert!(matches!(result, Err(Error::InvalidHttpRequestBody)));
-        }
-
-        #[test]
-        fn execute_fa_withdraw_succeeds() {
-            let mut host = MockHost::default();
-            let mut tx = Transaction::default();
-            let source = Address::User(jstz_mock::account1());
-
-            let ticket = TicketInfo {
-                id: 1234,
-                content: Some(b"random ticket content".to_vec()),
-                ticketer: jstz_mock::kt1_account1().into(),
-            }
-            .to_ticket(1)
-            .unwrap();
-
-            tx.begin();
-            TicketTable::add(&mut host, &mut tx, &source, &ticket.hash, 10).unwrap();
-            tx.commit(&mut host).unwrap();
-
-            let req = fa_withdraw_request();
-            let ticketer =
-                ContractKt1Hash::from_base58_check(jstz_mock::host::NATIVE_TICKETER)
-                    .unwrap();
-
-            execute(&mut host, &mut tx, &ticketer, &source, req)
-                .expect("Withdraw should not fail");
-
-            tx.begin();
-            assert_eq!(0, Account::balance(&host, &mut tx, &source).unwrap());
-
-            let level = host.run_level(|_| {});
-            assert_eq!(1, host.outbox_at(level).len());
-        }
-    }
-}
-
-pub mod deploy {
-    use super::*;
-    use crate::{operation, receipt};
-
-    pub fn deploy_smart_function(
-        hrt: &mut impl HostRuntime,
-        tx: &mut Transaction,
-        source: &impl Addressable,
-        function_code: ParsedCode,
-        account_credit: u64,
-    ) -> Result<SmartFunctionHash> {
-        let address = Account::create_smart_function(
-            hrt,
-            tx,
-            source,
-            account_credit,
-            function_code,
-        )?;
-        Account::sub_balance(hrt, tx, source, account_credit)?;
-        Ok(address)
-    }
-
-    pub fn execute(
-        hrt: &mut impl HostRuntime,
-        tx: &mut Transaction,
-        source: &impl Addressable,
-        deployment: operation::DeployFunction,
-    ) -> Result<receipt::DeployFunctionReceipt> {
-        let operation::DeployFunction {
-            function_code,
-            account_credit,
-        } = deployment;
-
-        // SAFETY: Smart function creation and sub_balance must be atomic
-        tx.begin();
-        match deploy_smart_function(hrt, tx, source, function_code, account_credit) {
-            Ok(address) => {
-                tx.commit(hrt)?;
-                debug_msg!(hrt, "[📜] Smart function deployed: {}\n", address);
-                Ok(receipt::DeployFunctionReceipt { address })
-            }
-            Err(err @ Error::AccountExists) => {
-                tx.rollback()?;
-                debug_msg!(hrt, "[📜] Smart function was already deployed\n");
-                Err(err)
-            }
-            Err(err) => {
-                tx.rollback()?;
-                debug_msg!(hrt, "[📜] Smart function deployment failed. \n");
-                Err(err)
-            }
-        }
-    }
-
-    #[cfg(test)]
-    mod test {
-        use crate::context::account::Address;
-
-        use super::*;
-        use jstz_core::kv::Transaction;
-        use jstz_mock::host::JstzMockHost;
-        use operation::DeployFunction;
-
-        #[test]
-        fn execute_deploy_deploys_smart_function_with_kt1_account1() {
-            let mut host = JstzMockHost::default();
-            let mut tx = Transaction::default();
-            let source = Address::User(jstz_mock::account1());
-            let hrt = host.rt();
-            tx.begin();
-
-            let deployment = DeployFunction {
-                function_code: "".to_string().try_into().unwrap(),
-                account_credit: 0,
-            };
-            let result = deploy::execute(hrt, &mut tx, &source, deployment);
-            assert!(result.is_ok());
-            let receipt = result;
-            assert!(receipt.is_ok());
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::str::FromStr;
-
-    use super::*;
-    use boa_engine::Context;
-    use http::header::{HeaderName, HeaderValue};
-    use jstz_api::http::header::HeadersClass;
-    use jstz_core::native::register_global_class;
-
-    fn create_test_request(headers: Vec<(String, String)>) -> JsResult<Request> {
-        let mut context = Context::default();
-        register_global_class::<RequestClass>(&mut context)?;
-        register_global_class::<HeadersClass>(&mut context)?;
-
-        let mut builder = http::Request::builder()
-            .method("POST")
-            .uri("jstz://test")
-            .body(Some(Vec::new()))
-            .map_err(|e| {
-                JsError::from_native(
-                    JsNativeError::error()
-                        .with_message(format!("Failed to create request: {}", e)),
-                )
-            })?;
-
-        // Set headers after building
-        let headers_map = builder.headers_mut();
-        for (key, value) in headers {
-            headers_map.insert(
-                HeaderName::from_str(&key).map_err(|e| {
-                    JsError::from_native(
-                        JsNativeError::error()
-                            .with_message(format!("Invalid header name: {}", e)),
-                    )
-                })?,
-                HeaderValue::from_str(&value).map_err(|e| {
-                    JsError::from_native(
-                        JsNativeError::error()
-                            .with_message(format!("Invalid header value: {}", e)),
-                    )
-                })?,
-            );
-        }
-
-        Request::from_http_request(builder, &mut context)
-    }
-
-    mod transfer_amount {
-        use super::*;
-        use std::ops::Deref;
-
-        struct TestRequest(Request);
-
-        impl Deref for TestRequest {
-            type Target = Request;
-
-            fn deref(&self) -> &Self::Target {
-                &self.0
-            }
-        }
-
-        fn wrap_request(request: Request) -> TestRequest {
-            TestRequest(request)
-        }
-
-        #[test]
-        fn test_valid_amount() -> JsResult<()> {
-            let request = wrap_request(create_test_request(vec![(
-                X_JSTZ_TRANSFER.to_string(),
-                "1000".to_string(),
-            )])?);
-            assert_eq!(
-                HostScript::extract_transfer_amount(&request.headers().deref())?,
-                Some(NonZeroU64::new(1000).unwrap())
-            );
-            Ok(())
-        }
-
-        #[test]
-        fn test_missing_header() -> JsResult<()> {
-            let request = wrap_request(create_test_request(vec![])?);
-            assert_eq!(
-                HostScript::extract_transfer_amount(&request.headers().deref())?,
-                None
-            );
-            Ok(())
-        }
-
-        #[test]
-        fn test_invalid_amount() -> JsResult<()> {
-            let request = wrap_request(create_test_request(vec![(
-                X_JSTZ_TRANSFER.to_string(),
-                "invalid".to_string(),
-            )])?);
-            assert!(
-                HostScript::extract_transfer_amount(&request.headers().deref()).is_err()
-            );
-            Ok(())
-        }
-    }
-}
